refactor(cargo-anvil): separate the anvil catalog from the reusable engine
Consolidate the base-catalog logic into a new anvil module and reduce the emit module to its generic drivers, removing the emit<->artifacts duplication.

- src/anvil/ now owns everything anvil-specific: the embedded templates, the artifacts registry (justfile/region/github/ado, each co-locating its constants with its builder functions), anvil_artifacts(), the Catalog::anvil() base catalog (an inherent impl so the catalog module stays anvil-free), and the plan-time reconcile helpers (Justfile case resolution, legacy migration, single-crate lint shape).

- src/catalog/ is now reusable-only: artifact, builder, meta. Catalog::from_parts keeps fields private while letting the anvil module assemble the base catalog.

- src/emit/ keeps only the generic plan_owned_file / plan_managed_region drivers; the per-area emitter files and dead aggregators are gone.

Public paths (artifacts::, Catalog::anvil) are unchanged; output is byte-identical (snapshot, fixture, extensibility, and schema tests pass).
