+++ crates/cargo-anvil/src/anvil/artifacts/ado.rs
@@ -1,0 +1,274 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+//! Azure DevOps Pipelines backend files: step templates, stages templates,
+//! and root pipelines, each an owned file gated on [`Backend::Ado`].
+//!
+//! Holds the embedded templates, the per-group fan-out, and the registry
+//! functions. See [`ado.md`](../../../docs/design/ado.md).
+
+use crate::backend::Backend;
+use crate::catalog::Artifact;
+
+/// Embedded body of the shared setup step template.
+pub const SETUP_STEP: &str = include_str!("../../../templates/ado/steps/setup.yml");
+
+/// Embedded body of the cargo-delta impact step template.
+pub const IMPACT_STEP: &str = include_str!("../../../templates/ado/steps/impact.yml");
+
+/// Embedded body of the advisory-comments step template.
+pub const ADVISORY_COMMENTS_STEP: &str = include_str!("../../../templates/ado/steps/advisory-comments.yml");
+
+/// Embedded body of the dirty-file job wrapper.
+pub const JOB_WRAPPER: &str = include_str!("../../../templates/ado/steps/job.yml");
+
+/// Embedded body of the PR-tier stages template.
+pub const PR_STAGES: &str = include_str!("../../../templates/ado/pr-stages.yml");
+
+/// Embedded body of the scheduled-tier stages template.
+pub const SCHEDULED_STAGES: &str = include_str!("../../../templates/ado/scheduled-stages.yml");
+
+/// Embedded body of the PR root pipeline.
+pub const PR_ROOT_PIPELINE: &str = include_str!("../../../templates/ado/pr-root-pipeline.yml");
+
+/// Embedded body of the scheduled root pipeline.
+pub const SCHEDULED_ROOT_PIPELINE: &str = include_str!("../../../templates/ado/scheduled-root-pipeline.yml");
+
+/// All check groups that get a per-group step template.
+///
+/// See [`super::github::GROUPS`] for the rationale around splitting
+/// `pr-slow` into three cloud-workflow-visible sub-stages.
+pub const GROUPS: &[&str] = &[
+    "pr-fast",
+    "pr-test",
+    "pr-runtime-analysis",
+    "pr-mutants",
+    "scheduled-test",
+    "scheduled-advisories",
+    "scheduled-exhaustive",
+];
+
+/// Embedded template for one per-group step. `__GROUP__` is substituted with
+/// the group name at emit time.
+pub const GROUP_STEP_TEMPLATE: &str = include_str!("../../../templates/ado/steps/group.yml");
+
+/// Placeholder token the per-group template uses for the group name.
+const GROUP_PLACEHOLDER: &str = "__GROUP__";
+
+/// Render the step template for one group.
+#[must_use]
+pub fn render_group_step(group: &str) -> String {
+    GROUP_STEP_TEMPLATE.replace(GROUP_PLACEHOLDER, group)
+}
+
+/// Repo-root-relative path for one group's step template.
+#[must_use]
+pub fn group_step_path(group: &str) -> String {
+    format!(".pipelines/anvil/steps/{group}.yml")
+}
+
+/// `.pipelines/anvil/steps/setup.yml`.
+#[must_use]
+pub fn setup_step() -> Artifact {
+    Artifact::backend_file(Backend::Ado, ".pipelines/anvil/steps/setup.yml", SETUP_STEP)
+}
+
+/// `.pipelines/anvil/steps/impact.yml`.
+#[must_use]
+pub fn impact_step() -> Artifact {
+    Artifact::backend_file(Backend::Ado, ".pipelines/anvil/steps/impact.yml", IMPACT_STEP)
+}
+
+/// `.pipelines/anvil/steps/advisory-comments.yml`.
+#[must_use]
+pub fn advisory_comments() -> Artifact {
+    Artifact::backend_file(Backend::Ado, ".pipelines/anvil/steps/advisory-comments.yml", ADVISORY_COMMENTS_STEP)
+}
+
+/// `.pipelines/anvil/steps/job.yml` — the dirty-file job wrapper.
+#[must_use]
+pub fn job_wrapper() -> Artifact {
+    Artifact::backend_file(Backend::Ado, ".pipelines/anvil/steps/job.yml", JOB_WRAPPER)
+}
+
+/// `.pipelines/anvil/pr.yml` — the PR-tier stages template.
+#[must_use]
+pub fn pr_stages() -> Artifact {
+    Artifact::backend_file(Backend::Ado, ".pipelines/anvil/pr.yml", PR_STAGES)
+}
+
+/// `.pipelines/anvil/scheduled.yml` — the scheduled-tier stages template.
+#[must_use]
+pub fn scheduled_stages() -> Artifact {
+    Artifact::backend_file(Backend::Ado, ".pipelines/anvil/scheduled.yml", SCHEDULED_STAGES)
+}
+
+/// `.pipelines/anvil-pr.yml` — the PR root pipeline.
+#[must_use]
+pub fn pr_root_pipeline() -> Artifact {
+    Artifact::backend_file(Backend::Ado, ".pipelines/anvil-pr.yml", PR_ROOT_PIPELINE)
+}
+
+/// `.pipelines/anvil-scheduled.yml` — the scheduled root pipeline.
+#[must_use]
+pub fn scheduled_root_pipeline() -> Artifact {
+    Artifact::backend_file(Backend::Ado, ".pipelines/anvil-scheduled.yml", SCHEDULED_ROOT_PIPELINE)
+}
+
+/// The per-group step templates, one concrete owned file per group.
+///
+/// Each `(group, path)` pair's `path` must equal [`group_step_path`] for its
+/// group (asserted in tests); the body is [`render_group_step`].
+pub(crate) const GROUP_STEPS: &[(&str, &str)] = &[
+    ("pr-fast", ".pipelines/anvil/steps/pr-fast.yml"),
+    ("pr-test", ".pipelines/anvil/steps/pr-test.yml"),
+    ("pr-runtime-analysis", ".pipelines/anvil/steps/pr-runtime-analysis.yml"),
+    ("pr-mutants", ".pipelines/anvil/steps/pr-mutants.yml"),
+    ("scheduled-test", ".pipelines/anvil/steps/scheduled-test.yml"),
+    ("scheduled-advisories", ".pipelines/anvil/steps/scheduled-advisories.yml"),
+    ("scheduled-exhaustive", ".pipelines/anvil/steps/scheduled-exhaustive.yml"),
+];
+
+/// All ADO backend artifacts in emission order.
+#[must_use]
+pub(crate) fn all() -> Vec<Artifact> {
+    let mut out = vec![setup_step(), impact_step(), advisory_comments(), job_wrapper()];
+    for (group, path) in GROUP_STEPS {
+        out.push(Artifact::backend_file(Backend::Ado, path, render_group_step(group)));
+    }
+    out.push(pr_stages());
+    out.push(scheduled_stages());
+    out.push(pr_root_pipeline());
+    out.push(scheduled_root_pipeline());
+    out
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn setup_and_impact_step_templates_are_non_empty() {
+        assert!(SETUP_STEP.contains("just anvil-setup"));
+        assert!(IMPACT_STEP.contains("cargo-delta"));
+        assert!(IMPACT_STEP.contains("##vso[task.setvariable"));
+    }
+
+    #[test]
+    fn setup_step_takes_group_parameter_and_dispatches() {
+        assert!(SETUP_STEP.contains("name: group"));
+        assert!(SETUP_STEP.contains("just anvil-setup"));
+        assert!(SETUP_STEP.contains("just anvil-${{ parameters.group }}-setup"));
+        assert!(SETUP_STEP.contains("eq(parameters.group, 'none')"));
+    }
+
+    #[test]
+    fn group_step_passes_group_to_setup() {
+        let body = render_group_step("pr-fast");
+        assert!(body.contains("template: setup.yml"));
+        assert!(body.contains("group: pr-fast"));
+    }
+
+    #[test]
+    fn impact_step_uses_group_none_and_installs_only_cargo_delta() {
+        assert!(IMPACT_STEP.contains("group: none"));
+        assert!(IMPACT_STEP.contains("anvil-tool-cargo-delta-install"));
+        assert!(!IMPACT_STEP.contains("cargo install --locked cargo-delta"));
+    }
+
+    #[test]
+    fn job_wrapper_declares_expected_contract() {
+        for needle in [
+            "name: name",
+            "name: pool",
+            "name: steps",
+            "type: stepList",
+            "name: artifacts",
+            "PublishPipelineArtifact@1",
+        ] {
+            assert!(JOB_WRAPPER.contains(needle), "wrapper missing '{needle}'");
+        }
+    }
+
+    #[test]
+    fn render_group_step_has_include_inputs_and_env() {
+        let body = render_group_step("pr-fast");
+        assert!(body.contains("parameters:"));
+        assert!(body.contains("name: include_modified"));
+        assert!(body.contains("name: include_affected"));
+        assert!(body.contains("name: include_required"));
+        assert!(body.contains("just anvil-pr-fast"));
+        assert!(body.contains("ANVIL_INCLUDE_MODIFIED"));
+        assert!(body.contains("ANVIL_INCLUDE_AFFECTED"));
+        assert!(body.contains("ANVIL_INCLUDE_REQUIRED"));
+        assert!(body.contains("PR_TITLE: $(System.PullRequest.Title)"));
+    }
+
+    #[test]
+    fn group_step_path_is_under_pipelines() {
+        assert_eq!(group_step_path("scheduled-test"), ".pipelines/anvil/steps/scheduled-test.yml");
+    }
+
+    #[test]
+    fn pr_stages_has_impact_and_group_stages() {
+        for needle in [
+            "stage: impact_linux",
+            "stage: impact_windows",
+            "stage: pr_fast",
+            "stage: pr_test",
+            "stage: pr_runtime_analysis",
+            "stage: pr_mutants",
+        ] {
+            assert!(PR_STAGES.contains(needle), "PR stages missing '{needle}'");
+        }
+        for needle in ["stage: pr_slow\n", "stage: pr_slow1\n", "stage: pr_slow2\n", "stage: pr_slow3\n"] {
+            assert!(
+                !PR_STAGES.contains(needle),
+                "Stale stage '{needle}' should be gone after the pr-slow rename"
+            );
+        }
+        assert!(PR_STAGES.contains("stageDependencies.impact_linux.compute.outputs"));
+        assert!(PR_STAGES.contains("stageDependencies.impact_windows.compute.outputs"));
+        assert!(PR_STAGES.contains("- template: steps/job.yml"));
+        assert!(
+            !PR_STAGES.contains("\n      - job: "),
+            "PR stages defines a bare `- job:` instead of going through steps/job.yml"
+        );
+        assert_eq!(
+            PR_STAGES.matches("- task: PublishCodeCoverageResults@2").count(),
+            2,
+            "cobertura publish should appear once per pr_test job (linux + windows)"
+        );
+    }
+
+    #[test]
+    fn scheduled_stages_has_three_groups() {
+        for needle in [
+            "stage: scheduled_test",
+            "stage: scheduled_advisories",
+            "stage: scheduled_exhaustive",
+        ] {
+            assert!(SCHEDULED_STAGES.contains(needle), "scheduled stages missing '{needle}'");
+        }
+        assert!(!SCHEDULED_STAGES.contains("scheduled_runtime"));
+        assert!(SCHEDULED_STAGES.contains("PublishCodeCoverageResults@2"));
+        assert!(SCHEDULED_STAGES.contains("- template: steps/job.yml"));
+        assert!(
+            !SCHEDULED_STAGES.contains("\n      - job: "),
+            "Scheduled stages defines a bare `- job:` instead of going through steps/job.yml"
+        );
+    }
+
+    #[test]
+    fn group_step_paths_match_render() {
+        assert_eq!(GROUP_STEPS.len(), GROUPS.len());
+        for ((group, path), expected_group) in GROUP_STEPS.iter().zip(GROUPS) {
+            assert_eq!(group, expected_group, "group order must match GROUPS");
+            assert_eq!(
+                *path,
+                group_step_path(group),
+                "registry path must match group_step_path for {group}"
+            );
+        }
+    }
+}
